requirements.txt
selenium==4.18.0
seleniumwire==5.1.0
requests==2.31.0
pyyaml==6.0.1
